% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simpleTables.R
\name{cox.table}
\alias{cox.table}
\title{Cox's Regression Table}
\usage{
cox.table(outcome, fu, predictors, varlabels)
}
\arguments{
\item{outcome}{The outcome variable. Should be a binary factor.}

\item{fu}{Time to event interval. Should be a numerical variable.}

\item{predictors}{A list of independent variables. They can include variable labels.}

\item{varlabels}{The variable labels. Should be a vector of labels.}
}
\description{
This function creates a Cox's proportional hazards regression table including OR, 95\% CI, and P values (unadjusted and adjusted using Hommel's correction).
}
\examples{
cox.table()
}
\keyword{cox,}
\keyword{regression}

