% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simpleTables.R
\name{descriptive.categorical.group}
\alias{descriptive.categorical.group}
\title{Comparison of Categorical Variables in a n-by-2 Contingency Table}
\usage{
descriptive.categorical.group(x, y)
}
\arguments{
\item{x}{A categorical (factor) variable}

\item{y}{A 2-level categorical (factor) variable}
}
\description{
This function compares 2 categorical variables in a n-by-2 contingency table.
}
\examples{
descriptive.categorical.group()
}
\keyword{factor,}
\keyword{numerical}

