% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simpleTables.R
\name{descriptive.categorical}
\alias{descriptive.categorical}
\title{Description of a Single Categorical Variable}
\usage{
descriptive.categorical(x)
}
\arguments{
\item{x}{A categorical (factor) variable}
}
\description{
This function describes a single categorical variable using a table of absolute and relative frequencies.
}
\examples{
descriptive.categorical()
}
\keyword{factor}

