% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simpleTables.R
\name{descriptive.numerical.group}
\alias{descriptive.numerical.group}
\title{Comparison of a Grouped Numerical Variable}
\usage{
descriptive.numerical.group(x, y)
}
\arguments{
\item{x}{A numerical variable}

\item{y}{The grouping (factor) variable}
}
\description{
This function describes a grouped numerical variable using a table with mean, SD, median, IQR, min, and max values by groups.
}
\examples{
descriptive.numerical.group()
}
\keyword{factor,}
\keyword{numerical}

