% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simpleTables.R
\name{descriptive.numerical}
\alias{descriptive.numerical}
\title{Description of a Single Numerical Variable}
\usage{
descriptive.numerical(x)
}
\arguments{
\item{x}{A numerical variable}
}
\description{
This function describes a single numerical variable using a table with mean, SD, median, IQR, min, and max values.
}
\examples{
descriptive.numerical()
}
\keyword{numerical}

