% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simplePlots.R
\name{logistic.plot}
\alias{logistic.plot}
\title{Forest plots for Logistic Regression Estimates}
\usage{
logistic.plot(outcome, predictors, varlabels)
}
\arguments{
\item{outcome}{The outcome variable. Should be a binary factor.}

\item{predictors}{A list of independent variables. They can include variable labels.}

\item{varlabels}{The variable labels. Should be a vector of labels.}
}
\description{
This function creates forest plots using OR and 95\% CI from a logistic regression.
}
\examples{
logistic.plot()
}
\keyword{forest,}
\keyword{logistic}

