% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simpleTables.R
\name{logistic.table}
\alias{logistic.table}
\title{Logistic Regression Table}
\usage{
logistic.table(outcome, predictors, varlabels)
}
\arguments{
\item{outcome}{The outcome variable. Should be a binary factor.}

\item{predictors}{A list of independent variables. They can include variable labels.}

\item{varlabels}{The variable labels. Should be a vector of labels.}
}
\description{
This function creates a logistic regression table including OR, 95\% CI, and P values (unadjusted and adjusted using Hommel's correction).
}
\examples{
logistic.table()
}
\keyword{logistic,}
\keyword{regression}

