% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simplePlots.R
\name{nice.colors}
\alias{nice.colors}
\title{Nice Colors for Plots}
\usage{
nice.colors(n)
}
\arguments{
\item{n}{Number of levels.}
}
\description{
This function creates the color pallete that will be used for plots.
}
\examples{
nice.colors()
}
\keyword{colors}

