% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simplePlots.R
\name{numerical.group.plot}
\alias{numerical.group.plot}
\title{Plots for Grouped Numerical Variables}
\usage{
numerical.group.plot(x, y, ...)
}
\arguments{
\item{x}{The numerical variable.}

\item{y}{The grouping factor variable.}

\item{...}{Other graphical parameters.}
}
\description{
This function describes grouped numerical variables using boxplots and applies the Kruskal-Wallis test.
}
\examples{
plot.numerical.group()
}
\keyword{factor,}
\keyword{numerical}

