% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simplePlots.R
\name{numerical.plot}
\alias{numerical.plot}
\title{Plots for Single Numerical Variable}
\usage{
numerical.plot(x, label)
}
\arguments{
\item{x}{A numerical variable.}

\item{label}{A label for the axes.}
}
\description{
This function describes single numerical variables using histograms and boxplots.
}
\examples{
plot.numerical()
}
\keyword{numerical}

