% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simplePlots.R
\name{plot.categorical.group}
\alias{plot.categorical.group}
\title{Plots for 2 or More Categorical Variables}
\usage{
\method{plot}{categorical.group}(x, y, align = "v", left = 4, ...)
}
\arguments{
\item{x}{The first factor variable.}

\item{y}{The second factor variable.}

\item{align}{Barplot alignment, either "v" (default) for vertical plots or "h" for horizontal plots.}

\item{...}{Other graphical parameters.}

\item{Left}{margin for horizontal plots. Default to 4.}
}
\description{
This function describes categorical variables from a n-by-n contingency table using juxtaposed bars and applies the Fisher's exact test.
}
\examples{
plot.categorical.group()
}
\keyword{factor}

