% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simplePlots.R
\name{plot.categorical}
\alias{plot.categorical}
\title{Plots for Single Categorical Variable}
\usage{
\method{plot}{categorical}(x, align = "v", left = 4, ...)
}
\arguments{
\item{x}{A factor variable.}

\item{align}{Barplot alignment, either "v" (default) for vertical plots or "h" for horizontal plots.}

\item{...}{Other graphical parameters.}

\item{Left}{margin for horizontal plots. Default to 4.}
}
\description{
This function describes categorical variables using (horizontal or vertical) barplots.
}
\examples{
plot.categorical()
}
\keyword{factor}

