% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simplePlots.R
\name{plot.numerical}
\alias{plot.numerical}
\title{Plots for Single Numerical Variable}
\usage{
\method{plot}{numerical}(x, label)
}
\arguments{
\item{x}{A numerical variable.}

\item{label}{A label for the axes.}
}
\description{
This function describes numerical variables using histograms and boxplots.
}
\examples{
plot.numerical()
}
\keyword{numerical}

