% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{simple}
\alias{simple}
\title{SimpleR}
\usage{
simple()
}
\value{

}
\description{
This function is for users with zero R experience who want to import a dataset,
view the dataset and variables, detect missing data, get summary statistics, remove/keep variables, rename variables,
change class of a variable, recode variable values and conduct an in-depth multiple regression with easy to understand
summaries.
}
\details{
This package contains one function simple that should be ran in the console to trigger and interactive experience where
the user will be prompted to import their dataset and then guided to conduct many data cleaning and analyzing tasks.
}
