% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simplePlots.R
\name{survival.plot}
\alias{survival.plot}
\title{Plots for Survival Curves}
\usage{
survival.plot(x, fu, outcome, title, position = "topright",
  logrank = "bottomleft", ...)
}
\arguments{
\item{x}{The grouping (predictor) variable.}

\item{fu}{Time to event interval. Should be a numeric variable.}

\item{outcome}{The outcome variable.}

\item{title}{Plot main title.}

\item{logrank}{Position of the legend showing the P value from the log-rank test. Default is "bottomleft".}

\item{...}{Other graphical parameters.}

\item{Position}{of the legend showing the levels of the predictor levels. Default is "topright".}
}
\description{
This function plots survival curves using the Kaplan-Meier method and compares them using the log-rank (Mantel-Cox) test.
}
\examples{
survival.plot()
}
\keyword{factor,}
\keyword{numerical,}
\keyword{survival}

