% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle.R
\name{tri_area}
\alias{tri_area}
\alias{tri_ix}
\alias{tri_jx}
\title{Area of triangles}
\usage{
tri_ix(x)

tri_jx(x)

tri_area(x)
}
\arguments{
\item{x}{}
}
\value{
area of triangles (a vector of length nrow(x)/3)
}
\description{
Input is x,y matrix in triplets
}
\details{
`tri_ix` builds the i-index
`tri_jx` builds the j-index
}
\examples{
set.seed(7)
lin <- cbind(1:10, sort(rnorm(10)))
tri <- assoc_triangle(lin)
a <- tri_area(lin[t(tri), ])
plot(lin)
scl <- function(x) (x - min(x))/diff(range(x))
col <- grey(seq(0.2, 0.8, length = 20), alpha = 0.8)[scl(a) * (19) + 1]
## smaller triangles are darker, get removed first
for (i in seq_len(nrow(tri))) polypath(lin[tri[i, ], ], col = col[i])
}
